i#3092 genapi: Split dr_modules.h from dr_tools.h
Multiple headers fed into dr_tools.h, which is no longer possible with
the simple copy-as-is export model.  We split the module types from
module_shared.h into a new file dr_modules.h, and move module routines
and types from instrument_api.h into the file as well.  The
implementations are also moved from instrument.c to a new file
module_api.c.

Issue: #3092
